update-mybatis-springboot对用户表的crud示例；

+++ pssms-rpcservice/src/main/java/me/geosmart/pssms/rpcs/domain/User.java
@@ -7,24 +7,24 @@ package me.geosmart.pssms.rpcs.domain;
  * @author Siva
  */
 public class User {
-    private Integer id;
+    private Long id;
     private String name;
     private String email;
 
     public User() {
     }
 
-    public User(Integer id, String name, String email) {
+    public User(Long id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
